requirements.txt
botframework-connector>=4.15.0
botbuilder-core
python-dotenv
pydantic
aiohttp
asyncpg
httpx
blibs
fastapi
uvicorn
asgi-logger
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-asyncpg
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-asyncio
opentelemetry-instrumentation-aiohttp-client
